requirements.txt
numpy==1.18.2
pandas==1.0.3
matplotlib==3.2.1
seaborn==0.10.0
hdx-python-api==4.3.9
-r docker-requirements.txt